refactor(sdk): extract network configs into relayer-configs module
Move config constants and mergeFhevmConfig out of relayer-utils into a
dedicated relayer-configs file to separate concerns.
